* updated mysql includes to 5.1 * added lib32 and lib64 directories for windows, added 64bit mysql libraries * added x64 configuration to vc2005 project (vc2003 doesn't support x64 :<) * 64bit compatibility fixes
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2041 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/BattlegroundMgr.h
@@ -380,9 +380,9 @@ public:
 	inline uint32 GetFreeSlots(uint32 t)
 	{
 		m_mainLock.Acquire();
-		uint32 s = m_playerCountPerTeam - m_players[t].size() - m_pendPlayers[t].size();
+		size_t s = m_playerCountPerTeam - m_players[t].size() - m_pendPlayers[t].size();
 		m_mainLock.Release();
-		return s;
+		return (uint32)s;
 	}
 
 	GameObject * SpawnGameObject(uint32 entry,uint32 MapId , float x, float y, float z, float o, uint32 flags, uint32 faction, float scale);
